refactor: move util codes to src
